requirements.txt
boto3>=1.16.9
botocore>=1.19.9
numpy>=1.19.3
pandas>=1.1.4
pytest>=6.1.2
psycopg2-binary>=2.8.6
Unidecode>=1.1.1
xlrd>=1.2.0
XlsxWriter>=1.3.7
FreqObjectOps>=0.1.5
